+++ app/scripts/views/list-search-view.js
@@ -97,6 +97,12 @@ var ListSearchView = Backbone.View.extend({
                 }
                 e.target.blur();
                 break;
+            case Keys.DOM_VK_A:
+                if (e.metaKey || e.ctrlKey) {
+                    e.stopPropagation();
+                    return;
+                }
+                return;
             default:
                 return;
         }