+++ app/scripts/models/entry-model.js
@@ -11,7 +11,7 @@ var EntryModel = Backbone.Model.extend({
     defaults: {},
     urlRegex: /^https?:\/\//i,
 
-    buildInFields: ['Title', 'Password', 'Notes', 'URL', 'UserName'],
+    builtInFields: ['Title', 'Password', 'Notes', 'URL', 'UserName'],
 
     initialize: function() {
     },
@@ -111,7 +111,7 @@ var EntryModel = Backbone.Model.extend({
     },
 
     _fieldsToModel: function(fields) {
-        return _.omit(fields, this.buildInFields);
+        return _.omit(fields, this.builtInFields);
     },
 
     _attachmentsToModel: function(binaries) {
@@ -142,8 +142,87 @@ var EntryModel = Backbone.Model.extend({
     matches: function(filter) {
         return !filter ||
             (!filter.tagLower || this.searchTags.indexOf(filter.tagLower) >= 0) &&
-            (!filter.textLower || this.searchText.indexOf(filter.textLower) >= 0) &&
+            (!filter.textLower || (filter.advanced ? this.matchesAdv(filter) : this.searchText.indexOf(filter.textLower) >= 0)) &&
             (!filter.color || filter.color === true && this.searchColor || this.searchColor === filter.color);
+    },
+
+    matchesAdv: function(filter) {
+        var adv = filter.advanced;
+        var search, match;
+        if (adv.regex) {
+            try { search = new RegExp(filter.text, adv.cs ? '' : 'i'); }
+            catch (e) { return false; }
+            match = this.matchRegex;
+        } else if (adv.cs) {
+            search = filter.text;
+            match = this.matchString;
+        } else {
+            search = filter.textLower;
+            match = this.matchStringLower;
+        }
+        if (this.matchEntry(this.entry, adv, match, search)) {
+            return true;
+        }
+        if (adv.history) {
+            for (var i = 0, len = this.entry.history.length; i < len; i++) {
+                if (this.matchEntry(this.entry.history[0], adv, match, search)) {
+                    return true;
+                }
+            }
+        }
+        return false;
+    },
+
+    matchString: function(str, find) {
+        return str.indexOf(find) >= 0;
+    },
+
+    matchStringLower: function(str, findLower) {
+        return str.toLowerCase().indexOf(findLower) >= 0;
+    },
+
+    matchRegex: function(str, regex) {
+        return regex.test(str);
+    },
+
+    matchEntry: function(entry, adv, compare, search) {
+        var matchField = this.matchField;
+        if (adv.user && matchField(entry, 'UserName', compare, search)) {
+            return true;
+        }
+        if (adv.url && matchField(entry, 'URL', compare, search)) {
+            return true;
+        }
+        if (adv.notes && matchField(entry, 'Notes', compare, search)) {
+            return true;
+        }
+        if (adv.pass && matchField(entry, 'Password', compare, search)) {
+            return true;
+        }
+        if (adv.other && matchField(entry, 'Title', compare, search)) {
+            return true;
+        }
+        var matches = false;
+        if (adv.other || adv.protect) {
+            var builtInFields = this.builtInFields;
+            var fieldNames = Object.keys(entry.fields);
+            matches = fieldNames.some(function (field) {
+                if (builtInFields.indexOf(field) >= 0) {
+                    return false;
+                }
+                if (typeof entry.fields[field] === 'string') {
+                    return adv.other && matchField(entry, field, compare, search);
+                } else {
+                    return adv.protect && matchField(entry, field, compare, search);
+                }
+            });
+        }
+        return matches;
+    },
+
+    matchField: function(entry, field, compare, search) {
+        var val = entry.fields[field];
+        return val ? compare(val.getText ? val.getText() : val, search) : false;
     },
 
     setColor: function(color) {
@@ -181,7 +260,7 @@ var EntryModel = Backbone.Model.extend({
     setField: function(field, val) {
         this._entryModified();
         var hasValue = val && (typeof val === 'string' || val instanceof kdbxweb.ProtectedValue && val.byteLength);
-        if (hasValue || this.buildInFields.indexOf(field) >= 0) {
+        if (hasValue || this.builtInFields.indexOf(field) >= 0) {
             this.entry.fields[field] = val;
         } else {
             delete this.entry.fields[field];