+++ app/scripts/views/fields/field-view-text.js
@@ -20,6 +20,8 @@ var FieldViewText = FieldView.extend({
 
     startEdit: function() {
         var text = this.getEditValue(this.value);
+        var isProtected = !!(this.value && this.value.isProtected);
+        this.$el.toggleClass('details__field--protected', isProtected);
         this.input = $(document.createElement(this.model.multiline ? 'textarea' : 'input'));
         this.valueEl.html('').append(this.input);
         this.input.attr({ autocomplete: 'off', spellcheck: 'false' })