+++ release-notes.md
@@ -7,6 +7,7 @@ Improvements
 `+` help/tips  
 `+` #50: notification on password copy  
 `-` #74: select all in search field  
+`*` #68: url display without http  
 
 ##### v0.5.1 (2015-12-15)
 Layout bugfixes